refactor(todel): clean up the `Conf` source code to be more maintainable (#93)
